Generate a simulated sparse matrix

Usage

simulate_sparse_matrix(
  nrow,
  ncol,
  sparsity = 0.95,
  distribution_fun = function(n) stats::rpois(n, lambda = 0.5) + 1,
  decimal = 0,
  seed = 1
)

Arguments

nrow

Number of rows (genes) in the matrix.

ncol

Number of columns (cells) in the matrix.

sparsity

Proportion of zero elements (sparsity level). Default is 0.95, meaning 95

distribution_funFunction to generate non-zero values.

decimalNumeric value, default is *`0`*. Controls the number of decimal places in the generated values. If set to *`0`*, values will be integers. When decimal > 0, random decimal parts are uniformly distributed across the full range.

seedRandom seed for reproducibility.

A sparse matrix of class "dgCMatrix" Generate a simulated sparse matrix simulate_sparse_matrix(1000, 500) |> check_sparsity()simulate_sparse_matrix(10, 10, decimal = 1) simulate_sparse_matrix(10, 10, decimal = 5)