Split indices.

An optimised version of split for the special case of splitting row indices into groups.

Usage

split_indices(group, n = 0L)

Arguments

group

Integer indices

n

The largest integer (may not appear in index). This is hint: if the largest value of group is bigger than n, the output will silently expand.

Value

A list of vectors of indices.

References

https://github.com/hadley/plyr/blob/d57f9377eb5d56107ba3136775f2f0f005f33aa3/src/split-numeric.cpp#L20

Examples

split_indices(sample(10, 100, rep = TRUE))
#> [[1]]
#> [1]  12  29  45  49  68  86  91  92 100
#> 
#> [[2]]
#>  [1]  4  6  8 34 50 63 83 89 90 98
#> 
#> [[3]]
#>  [1]  1 14 18 20 22 32 33 41 47 70 80 82 87 88
#> 
#> [[4]]
#>  [1] 10 17 30 38 39 51 69 73 75 78 93 97
#> 
#> [[5]]
#> [1] 24 28 53 54 58 76 79 81 96
#> 
#> [[6]]
#> [1]  2  3 15 26 62 65 71 74 85
#> 
#> [[7]]
#>  [1]  7 19 21 23 25 42 44 46 64 67 77
#> 
#> [[8]]
#> [1]  5 11 35 36 37 40 52 56 95
#> 
#> [[9]]
#>  [1]  9 13 16 31 43 55 59 61 72 94 99
#> 
#> [[10]]
#> [1] 27 48 57 60 66 84
#> 
split_indices(sample(10, 100, rep = TRUE), 10)
#> [[1]]
#>  [1] 26 29 36 40 45 61 66 78 82 87 88
#> 
#> [[2]]
#>  [1]  4 10 11 17 62 69 70 71 79 84 90 92
#> 
#> [[3]]
#> [1] 19 22 25 30 47 49 53
#> 
#> [[4]]
#>  [1] 13 39 48 55 59 64 77 89 93 95
#> 
#> [[5]]
#>  [1] 34 46 51 52 63 73 76 81 97 98
#> 
#> [[6]]
#>  [1]  2 20 27 33 41 44 56 67 80 86
#> 
#> [[7]]
#>  [1]  1  3 18 37 50 57 65 68 94 99
#> 
#> [[8]]
#> [1]  5  9 12 16 23 60 85 96
#> 
#> [[9]]
#>  [1]  6  7  8 14 15 21 24 31 38 42 54 58 74 83
#> 
#> [[10]]
#> [1]  28  32  35  43  72  75  91 100
#>