A sparse correlation function

Safe correlation function which returns a sparse matrix without missing values

Usage

sparse_cor(
  x,
  y = NULL,
  method = "pearson",
  allow_neg = TRUE,
  remove_na = TRUE,
  remove_inf = TRUE,
  ...
)

Arguments

x

Sparse matrix or character vector.

y

Sparse matrix or character vector.

method

Method to use for calculating the correlation coefficient.

allow_neg

Logical. Whether to allow negative values or set them to 0.

remove_na

Logical. Whether to replace NA values with 0.

remove_inf

Logical. Whether to replace infinite values with 1.

...

Other arguments passed to stats::cor function.

Value

A correlation matrix.

Examples

m1 <- simulate_sparse_matrix(
  500, 100
)
m2 <- simulate_sparse_matrix(
  500, 100,
  sparsity = 0.05
)
a <- sparse_cor(m1)
b <- sparse_cor(m1, m2)
c <- as_matrix(
  cor(as_matrix(m1)),
  sparse = TRUE
)
d <- as_matrix(
  cor(as_matrix(m1), as_matrix(m2)),
  sparse = TRUE
)

a[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>             col_1       col_2       col_3       col_4       col_5
#> col_1  1.00000000  0.03982146  0.03290085 -0.02022058  0.00827069
#> col_2  0.03982146  1.00000000 -0.04172518 -0.00276169 -0.03594182
#> col_3  0.03290085 -0.04172518  1.00000000  0.03481704 -0.03144989
#> col_4 -0.02022058 -0.00276169  0.03481704  1.00000000 -0.05034769
#> col_5  0.00827069 -0.03594182 -0.03144989 -0.05034769  1.00000000
c[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>             col_1       col_2       col_3       col_4       col_5
#> col_1  1.00000000  0.03982146  0.03290085 -0.02022058  0.00827069
#> col_2  0.03982146  1.00000000 -0.04172518 -0.00276169 -0.03594182
#> col_3  0.03290085 -0.04172518  1.00000000  0.03481704 -0.03144989
#> col_4 -0.02022058 -0.00276169  0.03481704  1.00000000 -0.05034769
#> col_5  0.00827069 -0.03594182 -0.03144989 -0.05034769  1.00000000
all.equal(a, c)
#> [1] "Attributes: < Component “i”: Numeric: lengths (5047, 5050) differ >"  
#> [2] "Attributes: < Component “p”: Mean relative difference: 0.0008422367 >"
#> [3] "Attributes: < Component “x”: Numeric: lengths (5047, 5050) differ >"  

b[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>             col_1       col_2         col_3       col_4        col_5
#> col_1  0.01405157 -0.01365356  0.0009438844 -0.01119988 -0.004190951
#> col_2 -0.04649233 -0.00733534  0.0312979324  0.02898365 -0.002140504
#> col_3  0.10292430 -0.03213596 -0.0439452327 -0.03683150 -0.029089599
#> col_4  0.01166789  0.03974206 -0.0393234755  0.05058941 -0.067618686
#> col_5 -0.05589691 -0.04225903  0.0512091815 -0.03204976  0.078861300
d[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>             col_1       col_2         col_3       col_4        col_5
#> col_1  0.01405157 -0.01365356  0.0009438844 -0.01119988 -0.004190951
#> col_2 -0.04649233 -0.00733534  0.0312979324  0.02898365 -0.002140504
#> col_3  0.10292430 -0.03213596 -0.0439452327 -0.03683150 -0.029089599
#> col_4  0.01166789  0.03974206 -0.0393234755  0.05058941 -0.067618686
#> col_5 -0.05589691 -0.04225903  0.0512091815 -0.03204976  0.078861300
all.equal(b, d)
#> [1] "Attributes: < Component “i”: Numeric: lengths (9996, 10000) differ >" 
#> [2] "Attributes: < Component “p”: Mean relative difference: 0.0004317249 >"
#> [3] "Attributes: < Component “x”: Numeric: lengths (9996, 10000) differ >" 

m1[sample(1:500, 10)] <- NA
m2[sample(1:500, 10)] <- NA

sparse_cor(m1, m2)[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>       col_1       col_2       col_3       col_4        col_5
#> col_1     .  .           .           .           .          
#> col_2     . -0.00733534  0.03129793  0.02898365 -0.002140504
#> col_3     . -0.03213596 -0.04394523 -0.03683150 -0.029089599
#> col_4     .  0.03974206 -0.03932348  0.05058941 -0.067618686
#> col_5     . -0.04225903  0.05120918 -0.03204976  0.078861300

system.time(
  sparse_cor(m1)
)
#>    user  system elapsed 
#>   0.003   0.000   0.003 
system.time(
  cor(as_matrix(m1))
)
#>    user  system elapsed 
#>   0.006   0.000   0.006 

system.time(
  sparse_cor(m1, m2)
)
#>    user  system elapsed 
#>   0.003   0.000   0.003 
system.time(
  cor(as_matrix(m1), as_matrix(m2))
)
#>    user  system elapsed 
#>   0.011   0.000   0.011