Split indices.

An optimised version of split for the special case of splitting row indices into groups.

Usage

split_indices(group, n = 0L)

Arguments

group

Integer indices

n

The largest integer (may not appear in index). This is hint: if the largest value of group is bigger than n, the output will silently expand.

Value

A list of vectors of indices.

References

https://github.com/hadley/plyr/blob/d57f9377eb5d56107ba3136775f2f0f005f33aa3/src/split-numeric.cpp#L20

Examples

split_indices(sample(10, 100, rep = TRUE))
#> [[1]]
#> [1]  4 10 40 46 58 64 73 76 78
#> 
#> [[2]]
#>  [1]  1  9 18 19 34 48 60 63 84 85 94 99
#> 
#> [[3]]
#> [1]   8  16  22  23  51  65  77  97 100
#> 
#> [[4]]
#> [1] 17 44 54 61 79 91
#> 
#> [[5]]
#> [1] 11 13 26 29 38 49 55 96
#> 
#> [[6]]
#> [1]  5 12 15 31 86 93
#> 
#> [[7]]
#>  [1]  2 14 24 25 27 32 33 35 47 52 53 74
#> 
#> [[8]]
#>  [1]  6 20 21 42 45 59 66 69 70 71 75 81 83 90 98
#> 
#> [[9]]
#>  [1]  7 28 36 37 39 43 57 62 67 68 80 87 88 89 95
#> 
#> [[10]]
#> [1]  3 30 41 50 56 72 82 92
#> 
split_indices(sample(10, 100, rep = TRUE), 10)
#> [[1]]
#> [1]  1  3 23 30 43 55 70 74
#> 
#> [[2]]
#> [1]  4  9 13 38 46 57 82 99
#> 
#> [[3]]
#>  [1]  2  5 17 26 59 61 62 68 69 85 94
#> 
#> [[4]]
#>  [1] 15 21 27 29 32 35 39 48 89 90 96
#> 
#> [[5]]
#>  [1] 12 14 24 25 41 42 45 47 64 67 80 81 88
#> 
#> [[6]]
#> [1] 11 40 60 73 79 84 92 98
#> 
#> [[7]]
#>  [1]  6 10 19 20 34 58 65 66 76 77 86
#> 
#> [[8]]
#>  [1]  7 16 22 28 36 44 52 53 63 75 91 95
#> 
#> [[9]]
#>  [1]   8  31  50  51  54  71  72  78  83  87  93  97 100
#> 
#> [[10]]
#> [1] 18 33 37 49 56
#>